apps/api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.7
httpx>=0.26.0
python-dotenv>=1.0.1
openai>=1.12.0
chromadb>=0.4.22
